% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{add_unique_id}
\alias{add_unique_id}
\title{Add unique identifiers to each DataSet in the provided DataSetList based on static attributes}
\usage{
add_unique_id(dsl, attrs)
}
\arguments{
\item{dsl}{The DataSetList}

\item{attrs}{The list of attributes to combine into a unique identifier}
}
\value{
A new DataSetList object where the split has been done based on the provided attributes, and the unique
identifier has been added.
}
\description{
Note that this function returns a new DataSetList object, since a split into new datasetlist has to be done to
ensure each dataset has exactly one unique identifier.
Note that only static attributes (see `get_static_attributes`) can be used to create unique identifiers.
}
\examples{
add_unique_id(dsl, c('instance'))
}
