% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{align_function_value}
\alias{align_function_value}
\title{Align data by function values}
\usage{
align_function_value(data, include_param = TRUE, format = IOHprofiler)
}
\arguments{
\item{data}{The data to align}

\item{include_param}{Whether to include the recorded parameters in the alignment}

\item{format}{Whether the data is form IOHprofiler or COCO.}
}
\value{
Data aligned by function value
}
\description{
Align data by function values
}
\keyword{internal}
