% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{arrange}
\alias{arrange}
\alias{arrange.DataSetList}
\title{S3 sort function for DataSetList}
\usage{
arrange(dsl, ...)

\method{arrange}{DataSetList}(dsl, ...)
}
\arguments{
\item{dsl}{The DataSetList to sort}

\item{...}{attribute by which `dsl` is sorted. Multiple attributes can be specified.}

\item{...}{attribute by which `dsl` is sorted. Multiple attributes can be specified.}
}
\description{
S3 sort function for DataSetList
}
\examples{
sort(dsl, DIM, -funcId, algId) 
}
