% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\name{as.character.DataSet}
\alias{as.character.DataSet}
\title{S3 generic as.character operator for DataSet}
\usage{
\method{as.character}{DataSet}(x, verbose = F, ...)
}
\arguments{
\item{x}{A DataSet object}

\item{verbose}{Verbose mode, currently not implemented}

\item{...}{Arguments passed to other methods}
}
\value{
A short description of the DataSet
}
\description{
S3 generic as.character operator for DataSet
}
