% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{AUC.ECDF}
\alias{AUC.ECDF}
\title{Area Under Curve (Empirical Cumulative Dsitribution Function)}
\usage{
\method{AUC}{ECDF}(fun, from = NULL, to = NULL)
}
\arguments{
\item{fun}{A ECDF object.}

\item{from}{double. Starting point of the area on x-axis}

\item{to}{double. Ending point of the area on x-axis}
}
\value{
a object of type 'ECDF'
}
\description{
Area Under Curve (Empirical Cumulative Dsitribution Function)
}
