% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\name{c.DataSet}
\alias{c.DataSet}
\title{S3 concatenation function for DataSet}
\usage{
\method{c}{DataSet}(...)
}
\arguments{
\item{...}{The DataSets to concatenate}
}
\value{
A new DataSet
}
\description{
Concatenation for DataSets. Combines multiple runs from separate DataSets
into a single DataSet object if all provided arguments have the same dimension, function ID and
algorithm ID, and each contains only a single run. Currently does not support parameter tracking
}
\examples{
c(dsl[[1]], dsl[[1]])
}
