% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{c.DataSetList}
\alias{c.DataSetList}
\title{S3 concatenation function for DataSetList}
\usage{
\method{c}{DataSetList}(...)
}
\arguments{
\item{...}{The DataSetLists to concatenate}
}
\value{
A new DataSetList
}
\description{
S3 concatenation function for DataSetList
}
\examples{
c(dsl[1], dsl[3])
}
