% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_align_runtime}
\alias{c_align_runtime}
\title{Align a dataSetList by runtime}
\usage{
c_align_runtime(data, FV, idxValue, maximization, idxTarget)
}
\arguments{
\item{data}{the data}

\item{FV}{Function values}

\item{idxValue}{index of the function values}

\item{maximization}{Boolean}

\item{idxTarget}{index of the target}
}
\description{
Align a dataSetList by runtime
}
