% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\name{cat.DataSet}
\alias{cat.DataSet}
\title{S3 generic cat operator for DataSet}
\usage{
cat.DataSet(x)
}
\arguments{
\item{x}{A DataSet object}
}
\value{
A short description of the DataSet
}
\description{
S3 generic cat operator for DataSet
}
\examples{
cat.DataSet(dsl[[1]])
}
