% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{clean_DataSetList}
\alias{clean_DataSetList}
\title{Clean DataSetList object by concatenating DataSets}
\usage{
clean_DataSetList(dsList)
}
\arguments{
\item{dsList}{The DataSetList object to clean}
}
\description{
Concatenates all DataSets with the same ID, function id and dimension
}
\examples{
clean_DataSetList(dsl)
}
