% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dsl}
\alias{dsl}
\title{Example DataSetList used in tests / examples}
\format{
DataSetList
}
\usage{
dsl
}
\description{
A DataSetList containing DataSets on 2 IOHProfiler functions from 2 algorithms in 16D
}
\examples{
summary(dsl)
}
\keyword{datasets}
