% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dsl_large}
\alias{dsl_large}
\title{Larger example DataSetList used in tests / examples}
\format{
DataSetList
}
\usage{
dsl_large
}
\description{
A DataSetList containing DataSets on all IOHProfiler functions from 11 algorithms in 100D
}
\examples{
summary(dsl_large)
}
\keyword{datasets}
