% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ECDF.DataSet}
\alias{ECDF.DataSet}
\title{Empirical Cumulative Dsitribution Function of Runtime of a single data set}
\usage{
\method{ECDF}{DataSet}(ds, ftarget)
}
\arguments{
\item{ds}{A DataSet object.}

\item{ftarget}{A Numerical vector. Function values at which runtime values are consumed}
}
\value{
a object of type 'ECDF'
}
\description{
Empirical Cumulative Dsitribution Function of Runtime of a single data set
}
