% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\name{==.DataSet}
\alias{==.DataSet}
\title{S3 generic == operator for DataSets}
\usage{
\method{==}{DataSet}(dsL, dsR)
}
\arguments{
\item{dsL}{A `DataSet` object}

\item{dsR}{A `DataSet` object}
}
\value{
True if the DataSets contain the same function, dimension and algorithm,
and have the exact same attributes
}
\description{
S3 generic == operator for DataSets
}
\examples{
dsl[[1]] == dsl[[2]]
}
