% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\name{fast_RT_samples}
\alias{fast_RT_samples}
\title{Function to get just the RT samples needed, without any formatting to improve speed}
\usage{
fast_RT_samples(RT_mat, target, maximization = F)
}
\arguments{
\item{RT_mat}{A matrix containing the RT-values of a dataset}

\item{target}{Which target-value to use}

\item{maximization}{Whether maximization is needed or not}
}
\description{
Function to get just the RT samples needed, without any formatting to improve speed
}
