% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{generate_data.AUC}
\alias{generate_data.AUC}
\title{Generate dataframe of a single function/dimension pair}
\usage{
generate_data.AUC(dsList, targets, which = "by_RT")
}
\arguments{
\item{dsList}{The DataSetList object}

\item{targets}{A list of the target value for which to calculate the AUC (Runtime or target value)}

\item{which}{Whether to use a fixed-target 'by_RT' perspective or fixed-budget 'by_FV'}
}
\description{
This function generates a dataframe which can be easily plotted using the `plot_general_data`-function
}
\examples{
generate_data.AUC(subset(dsl, funcId == 1), c(12, 16))
}
