% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{generate_data.EAF_Difference}
\alias{generate_data.EAF_Difference}
\title{Generate differences between two EAFs}
\usage{
generate_data.EAF_Difference(dsList1, dsList2)
}
\arguments{
\item{dsList1}{The first DataSetList object}

\item{dsList2}{The second DataSetList object}
}
\description{
This function uses the 'eaf' package to calculate eaf differences
}
\examples{
generate_data.EAF_Difference(dsl[1], dsl[3])
}
