% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{generate_data.ECDF_raw}
\alias{generate_data.ECDF_raw}
\title{Generate dataframe of a the unaggregated values of individual algorithms. Stripped-down version of}
\usage{
generate_data.ECDF_raw(dsList, targets, scale_log = F)
}
\arguments{
\item{dsList}{The DataSetList object}

\item{targets}{A list or data.table containing the targets per function / dimension. If this is
a data.table, it needs columns 'target', 'DIM' and 'funcId'}

\item{scale_log}{Wheterh to use logarithmic scaling or not}
}
\description{
This provides an unaggregated version of the function `generate_data.ECDF`.
}
\examples{
\dontshow{data.table::setDTthreads(1)}
generate_data.ECDF_raw(subset(dsl, funcId == 1), c(10, 15, 16))
}
