% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{generate_data.Heatmaps}
\alias{generate_data.Heatmaps}
\title{Nevergrad-dashboard based algorithm comparison}
\usage{
generate_data.Heatmaps(dsList, which = "by_FV", target_dt = NULL)
}
\arguments{
\item{dsList}{The DataSetList, can contain multiple functions and dimensions, but should have the
same algorithms for all of them. For functions/dimensions where this is not the case,
all algorithms are considered tied.}

\item{which}{Whether to use fixed-target ('by_FV') or fixed-budget ('by_RT') perspective}

\item{target_dt}{Custom data.table target value to use. When NULL, this is selected automatically.}
}
\value{
A matrix containing the pairwise win-ratios.
}
\description{
This procedure calculates the fraction of times algorithm A is better than
algorithm B according to their mean on each function,dimension,target tuple
}
\examples{
generate_data.Heatmaps(dsl)
generate_data.Heatmaps(dsl, which = 'by_RT')
}
