% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{generate_data.Parameter_correlation}
\alias{generate_data.Parameter_correlation}
\title{Generate dataframe of exactly 2 parameters, matched by running time}
\usage{
generate_data.Parameter_correlation(dsList, par1, par2)
}
\arguments{
\item{dsList}{The DataSetList object}

\item{par1}{The first parameter. Either a parameter name or 'f(x)'}

\item{par2}{The second parameter. Either a parameter name or 'f(x)'}
}
\description{
This function generates a dataframe which can be easily plotted using the `plot_general_data`-function
}
\examples{
generate_data.Parameter_correlation(subset(dsl, funcId == 1), 'f(x)', 'f(x)')
}
