% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{generate_data.Parameters}
\alias{generate_data.Parameters}
\title{Generate dataframe of a single function/dimension pair}
\usage{
generate_data.Parameters(dsList, which = "by_RT", scale_log = F)
}
\arguments{
\item{dsList}{The DataSetList object}

\item{which}{Whether to use a fixed-target 'by_RT' perspective or fixed-budget 'by_FV'}

\item{scale_log}{Wheterh to use logarithmic scaling or not}
}
\description{
This function generates a dataframe which can be easily plotted using the `plot_general_data`-function
}
\examples{
generate_data.Parameters(subset(dsl, funcId == 1))
}
