% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{generate_data.PMF}
\alias{generate_data.PMF}
\title{Generate dataframe of a single function/dimension pair for creating PDF or PMF plots}
\usage{
generate_data.PMF(dsList, target, which = "by_RT")
}
\arguments{
\item{dsList}{The DataSetList object}

\item{target}{The target value (Runtime or target value)}

\item{which}{Whether to use a fixed-target 'by_RT' perspective or fixed-budget 'by_FV'}
}
\description{
This function generates a dataframe which can be easily plotted using the `plot_general_data`-function
}
\examples{
generate_data.PMF(subset(dsl, funcId == 1), target = 15, which = 'by_RT')
}
