% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{get_AlgId}
\alias{get_AlgId}
\title{Get all algorithm ids present in a DataSetList}
\usage{
get_AlgId(data)
}
\arguments{
\item{data}{The DataSetLsit}
}
\value{
A sorted list of all unique algorithm ids which occur in the DataSetList
}
\description{
Get all algorithm ids present in a DataSetList
}
