% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{get_best_targets}
\alias{get_best_targets}
\title{Get the best function value reached in a DataSetList}
\usage{
get_best_targets(dsList, by = "funcId", maximize = T)
}
\arguments{
\item{dsList}{The DataSetLsit}
}
\value{
A list matrices of all runtime values which occur in the DataSetList
}
\description{
Get the best function value reached in a DataSetList
}
