% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_color_scheme}
\alias{get_color_scheme}
\title{Get colors according to the current colorScheme of the IOHanalyzer}
\usage{
get_color_scheme(algnames_in)
}
\arguments{
\item{algnames_in}{List of algorithms for which to get colors}
}
\description{
Get colors according to the current colorScheme of the IOHanalyzer
}
\examples{
get_color_scheme(get_algId(dsl))
}
