% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_color_scheme_dt}
\alias{get_color_scheme_dt}
\title{Get datatable of current color (and linestyle) scheme to file}
\usage{
get_color_scheme_dt()
}
\value{
data.table object with 3 columns: algnames, colors, linestyles
}
\description{
Get datatable of current color (and linestyle) scheme to file
}
\examples{
get_color_scheme_dt()
}
