% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_colorScheme}
\alias{get_colorScheme}
\title{Get colors according to the current colorScheme of the IOHanalyzer}
\usage{
get_colorScheme(n)
}
\arguments{
\item{n}{Number of colors to get}
}
\description{
Get colors according to the current colorScheme of the IOHanalyzer
}
\examples{
get_colorScheme(5)
}
