% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{get_DIM}
\alias{get_DIM}
\title{Get all dimensions present in a DataSetList}
\usage{
get_DIM(data)
}
\arguments{
\item{data}{The DataSetLsit}
}
\value{
A sorted list of all unique dimensions which occur in the DataSetList
}
\description{
Get all dimensions present in a DataSetList
}
