% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_dsc_omnibus}
\alias{get_dsc_omnibus}
\title{Perform omnibus statistical tests on the matrix of rankings from the DSCtool api}
\usage{
get_dsc_omnibus(res, method = NULL, alpha = 0.05)
}
\arguments{
\item{res}{The result of a call to the `get_dsc_rank`}

\item{method}{Which method to use to do the tests.
Has be be one of the allowed ones in `res$valid_methods`.
When NULL, the first valid option is chosen by default}

\item{alpha}{Threshold value for statistical significance}
}
\value{
A named list containing the algorithm means
}
\description{
Perform omnibus statistical tests on the matrix of rankings from the DSCtool api
}
\examples{
get_dsc_omnibus(get_dsc_rank(dsl))
}
