% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{get_ECDF_targets}
\alias{get_ECDF_targets}
\title{Generation of default ECDF-targets}
\usage{
get_ECDF_targets(dsList, type = "log-linear", number_targets = 10)
}
\arguments{
\item{dsList}{The DataSetList object for which to generate the targets}

\item{type}{The way to generate the targets. Either 'log-linear', 'linear' or 'bbob' (51 fixed targets,
equal for all functions / dimensions)}

\item{number_targets}{The amount of targets to generate}
}
\value{
A data.table with 3 columns: funcId, DIM and target
}
\description{
Generation of default ECDF-targets
}
\examples{
get_ECDF_targets(dsl, 'linear', 10)
}
