% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\name{get_ERT.DataSet}
\alias{get_ERT.DataSet}
\title{Get Expected RunTime}
\usage{
\method{get_ERT}{DataSet}(ds, ftarget)
}
\arguments{
\item{ds}{A DataSet object}

\item{ftarget}{The function target(s) for which to get the runtime summary}

\item{...}{Arguments passed to other methods}
}
\value{
A data.table containing the runtime statistics for each provided target
function value
}
\description{
Get Expected RunTime
}
