% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{get_ERT.DataSetList}
\alias{get_ERT.DataSetList}
\title{Get Expected RunTime}
\usage{
\method{get_ERT}{DataSetList}(ds, ftarget, algorithm = "all", ...)
}
\arguments{
\item{ds}{A DataSetList object}

\item{ftarget}{The function target(s) for which to get the ERT}

\item{algorithm}{Which algorithms in the DataSetList to consider.}

\item{...}{Arguments passed to other methods}
}
\value{
A data.table containing the runtime statistics for each provided target
function value
}
\description{
Get Expected RunTime
}
