% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\name{get_ERT}
\alias{get_ERT}
\title{Get Expected RunTime}
\usage{
get_ERT(ds, ...)
}
\arguments{
\item{ds}{A DataSet or DataSetList object}

\item{...}{Arguments passed to other methods}
}
\value{
A data.table containing the runtime samples for each provided target
function value
}
\description{
Get Expected RunTime
}
