% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{get_funcId}
\alias{get_funcId}
\title{Get all function ids present in a DataSetList}
\usage{
get_funcId(dsList)
}
\arguments{
\item{dsList}{The DataSetLsit}
}
\value{
A sorted list of all unique function ids which occur in the DataSetList
}
\description{
Get all function ids present in a DataSetList
}
