% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{get_funcName}
\alias{get_funcName}
\title{Get all function names present in a DataSetList}
\usage{
get_funcName(dsList)
}
\arguments{
\item{dsList}{The DataSetLsit}
}
\value{
A list of all unique function names which occur in the DataSetList
}
\description{
Get all function names present in a DataSetList
}
\examples{
get_funcName(dsl)
}
