% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{get_Funvals}
\alias{get_Funvals}
\title{Get all function values present in a DataSetList}
\usage{
get_Funvals(data)
}
\arguments{
\item{data}{The DataSetLsit}
}
\value{
A list matrices of all function values which occur in the DataSetList
}
\description{
Get all function values present in a DataSetList
}
