% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\name{get_FV}
\alias{get_FV}
\alias{get_FV.DataSet}
\title{Get function value matrix of the used dataset.}
\usage{
get_FV(ds, ...)

\method{get_FV}{DataSet}(ds, ...)
}
\arguments{
\item{ds}{The DataSet}

\item{...}{Arguments passed to other methods}
}
\value{
The matrix of FV values in the dataset, penalized if applicable.
}
\description{
To be used instead of accessing ds$FV directly, since in the case of constrained
problems, the violation handling should be applied before using the function values
Constraint penalty function should be set in global options, as IOHanalyzer.Violation_Function
}
\examples{
get_FV(dsl[[1]])
}
