% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\name{get_FV_overview.DataSet}
\alias{get_FV_overview.DataSet}
\title{Get Function Value condensed overview}
\usage{
\method{get_FV_overview}{DataSet}(ds, ...)
}
\arguments{
\item{ds}{A DataSet object}

\item{...}{Arguments passed to other methods}
}
\value{
A data.table containing the algorithm ID, best, worst and mean reached function
values, the number of runs and available budget for the DataSet
}
\description{
Get Function Value condensed overview
}
