% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\name{get_FV_overview}
\alias{get_FV_overview}
\title{Get Function Value condensed overview}
\usage{
get_FV_overview(ds, ...)
}
\arguments{
\item{ds}{A DataSet or DataSetList object}

\item{...}{Arguments passed to other methods}
}
\value{
A data.table containing the algorithm ID, best, worst and mean reached function
values, the number of runs and available budget for the DataSet
}
\description{
Get Function Value condensed overview
}
