% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\name{get_FV_RT_cross_sample}
\alias{get_FV_RT_cross_sample}
\title{Get Function Value and Time Cross Sample}
\usage{
get_FV_RT_cross_sample(ds, ...)
}
\arguments{
\item{ds}{A DataSet or DataSetList object}

\item{...}{Arguments passed to other methods}
}
\value{
A data.table containing the function runs amount statistics for each provided
target runtime and target value
}
\description{
Get Function Value and Time Cross Sample
}
