% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\name{get_FV_sample.DataSet}
\alias{get_FV_sample.DataSet}
\title{Get Funtion Value Samples}
\usage{
\method{get_FV_sample}{DataSet}(ds, runtime, output = "wide", ...)
}
\arguments{
\item{ds}{A DataSet object}

\item{runtime}{A Numerical vector. Runtimes at which function values are reached}

\item{output}{A character. The format of the output data: 'wide' or 'long'}

\item{...}{Arguments passed to other methods}
}
\value{
A data.table containing the function value samples for each provided
target runtime
}
\description{
Get Funtion Value Samples
}
