% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R, R/DataSetList.R
\name{get_FV_sample}
\alias{get_FV_sample}
\alias{get_FV_sample.DataSet}
\alias{get_FV_sample.DataSetList}
\title{Get Funtion Value Samples}
\usage{
get_FV_sample(ds, ...)

\method{get_FV_sample}{DataSet}(ds, runtime, output = "wide", ...)

\method{get_FV_sample}{DataSetList}(ds, runtime, algorithm = "all", ...)
}
\arguments{
\item{ds}{A DataSet or DataSetList object}

\item{...}{Arguments passed to other methods}

\item{runtime}{A Numerical vector. Runtimes at which function values are reached}

\item{output}{A String. The format of the output data: 'wide' or 'long'}

\item{algorithm}{Which algorithms in the DataSetList to consider.}
}
\value{
A data.table containing the function value samples for each provided
target runtime
}
\description{
Get Funtion Value Samples
}
\examples{
get_FV_sample(dsl, 100)
get_FV_sample(dsl[[1]], 100)
}
