% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{get_FV_summary.DataSetList}
\alias{get_FV_summary.DataSetList}
\title{Get Function Value Summary}
\usage{
\method{get_FV_summary}{DataSetList}(ds, runtime, algorithm = "all", ...)
}
\arguments{
\item{ds}{A DataSetList object}

\item{runtime}{A Numerical vector. Runtimes at which function values are reached}

\item{algorithm}{Which algorithms in the DataSetList to consider.}

\item{...}{Arguments passed to other methods}
}
\value{
A data.table containing the function value statistics for each provided
target runtime value
}
\description{
Get Function Value Summary
}
