% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\name{get_FV_summary}
\alias{get_FV_summary}
\title{Get Function Value Summary}
\usage{
get_FV_summary(ds, ...)
}
\arguments{
\item{ds}{A DataSet or DataSetList object}

\item{...}{Arguments passed to other methods}
}
\value{
A data.table containing the function value statistics for each provided
target runtime value
}
\description{
Get Function Value Summary
}
