% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{get_id}
\alias{get_id}
\title{Get the unique identifiers for each DataSet in the provided DataSetList}
\usage{
get_id(dsl)
}
\arguments{
\item{dsl}{The DataSetList}
}
\value{
The list of unique identiefiers present in dsl
}
\description{
If no unique identifier is set (using `change_id` or done in DataSet construction from 1.6.0 onwards), 
this function falls back on returning the algorith id (from `get_aldId`)to ensure backwards compatibility
}
\examples{
get_id(dsl)
}
