% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_line_style}
\alias{get_line_style}
\title{Get line styles according to the current styleScheme of the IOHanalyzer}
\usage{
get_line_style(ids_in)
}
\arguments{
\item{ids_in}{List of algorithms (or custom ids, see `change_id`) for which to get linestyles}
}
\description{
Get line styles according to the current styleScheme of the IOHanalyzer
}
\examples{
get_line_style(get_algId(dsl))
}
