% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R, R/DataSetList.R
\name{get_maxRT}
\alias{get_maxRT}
\alias{get_maxRT.DataSet}
\alias{get_maxRT.DataSetList}
\title{Get the maximal running time}
\usage{
get_maxRT(ds, ...)

\method{get_maxRT}{DataSet}(ds, output = "wide", ...)

\method{get_maxRT}{DataSetList}(ds, algorithm = "all", ...)
}
\arguments{
\item{ds}{A DataSet or DataSetList object}

\item{...}{Arguments passed to other methods}

\item{output}{The format of the outputted table: 'wide' or 'long'}

\item{algorithm}{DEPRECATED, will be removed in next release. Which algorithms in the DataSetList to consider.}
}
\value{
A data.table object containing the algorithm ID and the running time
when the algorithm terminates in each run
}
\description{
Get the maximal running time
}
\examples{
get_maxRT(dsl)
get_maxRT(dsl[[1]])
}
