% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_ontology_data}
\alias{get_ontology_data}
\title{Get the list of available options for data from the OPTION ontology}
\usage{
get_ontology_data(datasource, fids, dims, algs, iids = NULL,
  funcsuites = NULL)
}
\arguments{
\item{datasource}{The datasource for which to get the attributes. Either BBOB or Nevergrad}

\item{varname}{The variable for which to get the options. Restricted to [Fid, Iid, DIM, AlgId, Suite]}
}
\value{
the options of varname given the specified datasource
}
\description{
Get the list of available options for data from the OPTION ontology
}
\examples{
get_ontology_data("BBOB", 5, 5, "IPOP400D", 1)
}
