% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R, R/DataSetList.R
\name{get_overview}
\alias{get_overview}
\alias{get_overview.DataSet}
\alias{get_overview.DataSetList}
\title{Get condensed overview of datasets}
\usage{
get_overview(ds, ...)

\method{get_overview}{DataSet}(ds, ...)

\method{get_overview}{DataSetList}(ds, ...)
}
\arguments{
\item{ds}{A DataSet or DataSetList object}

\item{...}{Arguments passed to other methods}
}
\value{
A data.table containing some basic information about the provided DataSet(List)
}
\description{
Get condensed overview of datasets
}
\examples{
get_overview(dsl)
get_overview(dsl[[1]])
}
