% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\name{get_PAR_sample.DataSet}
\alias{get_PAR_sample.DataSet}
\title{Get Parameter Value Samples}
\usage{
\method{get_PAR_sample}{DataSet}(ds, ftarget, parId = "all",
  output = "wide", ...)
}
\arguments{
\item{ds}{A DataSet object}

\item{ftarget}{A Numerical vector. Function values at which parameter values are observed}

\item{parId}{A character vector. Either 'all' or the name of parameters to be retrieved}

\item{output}{A character. The format of the output data: 'wide' or 'long'}

\item{...}{Arguments passed to other methods}
}
\value{
A data.table object containing parameter values aligned at each given target value
}
\description{
Get Parameter Value Samples
}
