% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\name{get_PAR_summary.DataSet}
\alias{get_PAR_summary.DataSet}
\title{Get Parameter Value Summary}
\usage{
\method{get_PAR_summary}{DataSet}(ds, ftarget, parId = "all", ...)
}
\arguments{
\item{ds}{A DataSet object}

\item{ftarget}{A Numerical vector. Function values at which parameter values are observed}

\item{parId}{A character vector. Either 'all' or the name of parameters to be retrieved}

\item{...}{Arguments passed to other methods}
}
\value{
A data.table object containing basic statistics of parameter values aligned
at each given target value
}
\description{
Get Parameter Value Summary
}
