% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{get_PAR_summary.DataSetList}
\alias{get_PAR_summary.DataSetList}
\title{Get Parameter Value Summary}
\usage{
\method{get_PAR_summary}{DataSetList}(ds, ftarget, algorithm = "all",
  ...)
}
\arguments{
\item{ds}{A DataSetList object}

\item{ftarget}{A Numerical vector. Function values at which parameter values are observed}

\item{algorithm}{Which algorithms in the DataSetList to consider.}

\item{...}{Arguments passed to other methods}
}
\value{
A data.table object containing basic statistics of all parameter values aligned
at each given target value
}
\description{
Get Parameter Value Summary
}
