% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R, R/DataSetList.R
\name{get_PAR_summary}
\alias{get_PAR_summary}
\alias{get_PAR_summary.DataSet}
\alias{get_PAR_summary.DataSetList}
\title{Get Parameter Value Summary}
\usage{
get_PAR_summary(ds, ftarget, ...)

\method{get_PAR_summary}{DataSet}(ds, ftarget, parId = "all", ...)

\method{get_PAR_summary}{DataSetList}(ds, ftarget, algorithm = "all",
  ...)
}
\arguments{
\item{ds}{A DataSet or DataSetList object}

\item{ftarget}{A Numerical vector. Function values at which parameter values are observed}

\item{...}{Arguments passed to other methods}

\item{parId}{A character vector. Either 'all' or the name of parameters to be retrieved}

\item{algorithm}{Which algorithms in the DataSetList to consider.}
}
\value{
A data.table object containing basic statistics of parameter values aligned at each given target value
}
\description{
Get Parameter Value Summary
}
\examples{
get_PAR_summary(dsl, 14)
get_PAR_summary(dsl[[1]], 14)
}
