% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{get_position_dsl}
\alias{get_position_dsl}
\title{Extract the position information from a datasetlist object}
\usage{
get_position_dsl(dsList, iid)
}
\arguments{
\item{dsList}{The DataSetList object}

\item{iid}{the Instance Id from which to get the position history (can be a list)}
}
\description{
Extract the position information from a datasetlist object
}
\examples{
get_position_dsl(subset(dsl, funcId == 1), 1)
}
