% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOHanalyzer.R
\name{get_property}
\alias{get_property}
\title{Get the value of some global IOHanalyzer properties}
\usage{
get_property(prop_name, ...)
}
\arguments{
\item{prop_name}{The property to get. Current options are:
\itemize{
\item 'probs': The probabilities in the displayed quantiles (RT and FV summaries).
\item 'max_samples': The maximum number of samples to generate for each algorithm.
}}

\item{...}{Arguments passed to underlying functions (i.e. number of colors to get colorscheme for)}
}
\description{
Get the value of some global IOHanalyzer properties
}
\examples{
get_property("probs")
}
