% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\name{get_RT}
\alias{get_RT}
\alias{get_RT.DataSet}
\title{Get runtime matrix of the used dataset.}
\usage{
get_RT(ds, ...)

\method{get_RT}{DataSet}(ds, ...)
}
\arguments{
\item{ds}{The DataSet}

\item{...}{Arguments passed to other methods}
}
\value{
The matrix of FV values in the dataset, penalized if applicable.
}
\description{
To be used instead of accessing ds$RT directly, since in the case of constrained
problems, the violation handling should be applied before using the function values
Constraint penalty function should be set in global options, as IOHanalyzer.Violation_Function
}
\examples{
get_RT(dsl[[1]])
}
