% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{get_RT_overview.DataSetList}
\alias{get_RT_overview.DataSetList}
\title{Get Runtime Value condensed overview}
\usage{
\method{get_RT_overview}{DataSetList}(ds, algorithm = "all", ...)
}
\arguments{
\item{ds}{A DataSetList object}

\item{algorithm}{Which algorithms in the DataSetList to consider.}

\item{...}{Arguments passed to other methods}
}
\value{
A data.table containing the algorithm ID, minimum and maximum used evaluations,
number of runs and available budget for the DataSet
}
\description{
Get Runtime Value condensed overview
}
