% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\name{get_RT_overview}
\alias{get_RT_overview}
\title{Get Runtime Value condensed overview}
\usage{
get_RT_overview(ds, ...)
}
\arguments{
\item{ds}{A DataSet or DataSetList object}

\item{...}{Arguments passed to other methods}
}
\value{
A data.table containing the algorithm ID, minimum and maximum used evaluations,
number of runs and available budget for the DataSet
}
\description{
Get Runtime Value condensed overview
}
