% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\name{get_RT_sample.DataSet}
\alias{get_RT_sample.DataSet}
\title{Get RunTime Sample}
\usage{
\method{get_RT_sample}{DataSet}(ds, ftarget, output = "wide", ...)
}
\arguments{
\item{ds}{A DataSet object}

\item{ftarget}{A Numerical vector. Function values at which runtime values are consumed}

\item{output}{A character determining the format of output data.table: 'wide' or 'long'}

\item{...}{Arguments passed to other methods}
}
\value{
A data.table containing the runtime samples for each provided target
function value
}
\description{
Get RunTime Sample
}
