% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{get_RT_sample.DataSetList}
\alias{get_RT_sample.DataSetList}
\title{Get RunTime Sample}
\usage{
\method{get_RT_sample}{DataSetList}(ds, ftarget, algorithm = "all", ...)
}
\arguments{
\item{ds}{A DataSetList object}

\item{ftarget}{A Numerical vector. Function values at which runtime values are consumed}

\item{algorithm}{Which algorithms in the DataSetList to consider.}

\item{...}{Arguments passed to other methods}
}
\value{
A data.table containing the runtime samples for each provided target
function value
}
\description{
Get RunTime Sample
}
