% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\name{get_RT_sample}
\alias{get_RT_sample}
\title{Get RunTime Sample}
\usage{
get_RT_sample(ds, ...)
}
\arguments{
\item{ds}{A DataSet or DataSetList object}

\item{...}{Arguments passed to other methods}
}
\value{
A data.table containing the runtime samples for each provided target
function value
}
\description{
Get RunTime Sample
}
