% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\name{get_RT_summary.DataSet}
\alias{get_RT_summary.DataSet}
\title{Get RunTime Summary}
\usage{
\method{get_RT_summary}{DataSet}(ds, ftarget, ...)
}
\arguments{
\item{ds}{A DataSet object}

\item{ftarget}{The function target(s) for which to get the runtime summary}

\item{...}{Arguments passed to other methods}
}
\value{
A data.table containing the runtime statistics for each provided target
function value
}
\description{
Get RunTime Summary
}
