% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\name{get_RT_summary}
\alias{get_RT_summary}
\title{Get RunTime Summary}
\usage{
get_RT_summary(ds, ...)
}
\arguments{
\item{ds}{A DataSet or DataSetList object}

\item{...}{Arguments passed to other methods}
}
\value{
A data.table containing the runtime statistics for each provided target
function value
}
\description{
Get RunTime Summary
}
