% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{get_runtimes}
\alias{get_runtimes}
\title{Get all runtime values present in a DataSetList}
\usage{
get_runtimes(dsList)
}
\arguments{
\item{dsList}{The DataSetLsit}
}
\value{
A list matrices of all runtime values which occur in the DataSetList
}
\description{
Get all runtime values present in a DataSetList
}
