% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{get_static_attribute_values}
\alias{get_static_attribute_values}
\title{Get all options for a specific attribute which can be used to subset a DataSetList}
\usage{
get_static_attribute_values(dsl, attribute)
}
\arguments{
\item{dsl}{The DataSetList}

\item{attribute}{the name of the attribute for which to get the available options in dsl}
}
\value{
The list of options for the specified attribute
}
\description{
This is a more generic version of the existing `get_dim`, `get_funcId` and `get_algId` functions.
Note the only attributes returned by `get_static_attributes` are supported in this funcion
}
\examples{
get_static_attribute_values(dsl, 'funcId')
}
