% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{get_static_attributes}
\alias{get_static_attributes}
\title{Get all attributes which can be used to subset a DataSetList}
\usage{
get_static_attributes(dsl)
}
\arguments{
\item{dsl}{The DataSetList}
}
\value{
The list of available attributes
}
\description{
Get all attributes which can be used to subset a DataSetList
}
\examples{
get_static_attributes(dsl)
}
