% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_target_dt}
\alias{get_target_dt}
\title{Generate datatables of runtime or function value targets for a DataSetList}
\usage{
get_target_dt(dsList, which = "by_RT")
}
\arguments{
\item{dsList}{A DataSetList}

\item{which}{Whether to generate fixed-target ('by_FV') or fixed-budget ('by_RT') targets}
}
\value{
a data.table of targets
}
\description{
Only one target is generated per (function, dimension)-pair, as opposed to the
function `get_default_ECDF_targets`, which generates multiple targets.
}
\examples{
get_target_dt(dsl)
}
