% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{get_unique_id}
\alias{get_unique_id}
\title{Get the unique identifiers for each DataSet in the provided DataSetList}
\usage{
get_unique_id(dsl)
}
\arguments{
\item{dsl}{The DataSetList}
}
\value{
The list of unique identiefiers present in dsl
}
\description{
If no unique identifier is set (using `add_unique_id`), this function falls back on returning the algorith id
(from `get_aldId`)to ensure backwards compatibility
}
\examples{
get_unique_id(dsl)
}
