% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{IOH_plot_ly_default}
\alias{IOH_plot_ly_default}
\title{Template for creating plots in the IOHanalyzer-style}
\usage{
IOH_plot_ly_default(title = NULL, x.title = NULL, y.title = NULL)
}
\arguments{
\item{title}{Title for the plot}

\item{x.title}{X-axis label}

\item{y.title}{Y-axis label}
}
\description{
Template for creating plots in the IOHanalyzer-style
}
\examples{
IOH_plot_ly_default("Example plot","x-axis","y-axis")
}
