% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOHanalyzer-deprecated.R
\name{IOHanlyzer-deprecated}
\alias{IOHanlyzer-deprecated}
\title{Deprecated function in package \pkg{IOHanalyzer}}
\description{
The functions listed below are deprecated and will be defunct in
  the near future. When possible, alternative functions with similar
  functionality are also mentioned.
}
\keyword{internal}
