% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{limit.data}
\alias{limit.data}
\title{Reduce the size of the data set by evenly subsampling the records}
\usage{
limit.data(df, n)
}
\arguments{
\item{df}{The data to subsample}

\item{n}{The amount of samples}
}
\value{
A smaller data.frame
}
\description{
Reduce the size of the data set by evenly subsampling the records
}
