% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{max_ERTs.DataSetList}
\alias{max_ERTs.DataSetList}
\title{S3 generic function to get the ERT-values for all DataSets in a DataSetList at certain targets}
\usage{
\method{max_ERTs}{DataSetList}(dsList, aggr_on = "funcId",
  targets = NULL, maximize = T)
}
\arguments{
\item{dsList}{The DataSetLsit}

\item{aggr_on}{Whether to aggregate on 'funcId' or 'DIM'.}

\item{targets}{Predifined target function-values. Should be one for each function/dimension}

\item{maximize}{Whether the DataSetList is from a maximization or minimization problem}
}
\value{
A data.table containing ERT-values
}
\description{
S3 generic function to get the ERT-values for all DataSets in a DataSetList at certain targets
}
