% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{mean_FVs.DataSetList}
\alias{mean_FVs.DataSetList}
\title{S3 generic function to get the expected function-values for all DataSets in a
DataSetList at certain runtimes}
\usage{
\method{mean_FVs}{DataSetList}(dsList, aggr_on = "funcId",
  runtimes = NULL)
}
\arguments{
\item{dsList}{The DataSetLsit}

\item{aggr_on}{Whether to aggregate on 'funcId' or 'DIM'.}

\item{runtimes}{Predifined target runtimes-values. Should be one for each function/dimension}
}
\value{
A data.table containing expected fucntion-values
}
\description{
S3 generic function to get the expected function-values for all DataSets in a
DataSetList at certain runtimes
}
