% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOHanalyzer-deprecated.R
\name{mean_FVs}
\alias{mean_FVs}
\alias{mean_FVs.DataSetList}
\title{Get the expected function-values for all DataSets in a DataSetList at certain runtimes}
\usage{
mean_FVs(dsList, aggr_on = "funcId", runtimes = NULL)

\method{mean_FVs}{DataSetList}(dsList, aggr_on = "funcId", runtimes = NULL)
}
\arguments{
\item{dsList}{The DataSetLsit}

\item{aggr_on}{Whether to aggregate on 'funcId' or 'DIM'.}

\item{runtimes}{Predifined target runtimes-values. Should be one for each function/dimension}
}
\value{
A data.table containing expected fucntion-values
}
\description{
Get the expected function-values for all DataSets in a DataSetList at certain runtimes
}
\examples{
mean_FVs(dsl)
}
