% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{pairwise.test}
\alias{pairwise.test}
\title{Performs a pairwise statistical (ks-) test on the bootstrapped hitting times of a datasetlist}
\usage{
pairwise.test(dsl, ftarget, alpha = 0.01, bootstrap.size = 30)
}
\arguments{
\item{dsl}{A datasetlist, consisiting of at least 2 algorithms on a single function and dimension}

\item{ftarget}{The target value for calculating the hitting times}

\item{alpha}{The cutoff for statistical significance}

\item{bootstrap.size}{The maximum bootstrap size}
}
\value{
A matrix containing p-values indicating whether or not the ROW is statistically significantly
better than the COLUMN
}
\description{
Performs a pairwise statistical (ks-) test on the bootstrapped hitting times of a datasetlist
}
\examples{
pairwise.test(subset(dsl, funcId==1), 16)
}
