% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDataSetList.R
\name{Plot.BiObj.Difference}
\alias{Plot.BiObj.Difference}
\alias{Plot.BiObj.Difference.DataSetList}
\title{Create a candlestick plot of Glicko2-rankings}
\usage{
Plot.BiObj.Difference(
  dsList,
  xmin = NULL,
  xmax = NULL,
  ymin = NULL,
  ymax = NULL,
  scale.xlog = F,
  scale.ylog = F,
  show.min = T,
  show.max = T
)

\method{Plot.BiObj.Difference}{DataSetList}(
  dsList,
  xmin = NULL,
  xmax = NULL,
  ymin = NULL,
  ymax = NULL,
  scale.xlog = F,
  scale.ylog = F,
  show.min = T,
  show.max = T
)
}
\arguments{
\item{dsList}{A DataSetList}

\item{xmin}{The starting runtime (optional)}

\item{xmax}{The final runtime (optional)}

\item{ymin}{The minimal function value (optional)}

\item{ymax}{The maximal function value (optional)}

\item{scale.xlog}{Whether or not to scale the x-axis logaritmically}

\item{scale.ylog}{Whether or not to scale the y-axis logaritmically}

\item{show.min}{Whether or not to show the outline of the minimum reached
function values across the considered algorithms}

\item{show.max}{Whether or not to show the outline of the maximum reached
function values across the considered algorithms}
}
\description{
Create a candlestick plot of Glicko2-rankings
}
\examples{
Plot.BiObj.Difference(subset(dsl, funcId == 2))
}
