% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDataSetList.R
\name{Plot.Comparison.Heatmap}
\alias{Plot.Comparison.Heatmap}
\alias{Plot.Comparison.Heatmap.DataSetList}
\title{Plot a heatmap according to the specifications from the Nevergrad dashboard}
\usage{
Plot.Comparison.Heatmap(dsList, target_dt, which = "by_FV")

\method{Plot.Comparison.Heatmap}{DataSetList}(dsList, target_dt = NULL, which = "by_FV")
}
\arguments{
\item{dsList}{A DataSetList (should consist of only one function and dimension).}

\item{target_dt}{A data-table containing the targets to condider on each function/dimension pair}

\item{which}{Whether to use fixed-target ('by_FV') or fixed-budget ('by_RT') perspective}
}
\value{
A heatmap showing the fraction of times algorithm A beats algorithm B
}
\description{
Plot a heatmap according to the specifications from the Nevergrad dashboard
}
\examples{
Plot.Comparison.Heatmap(dsl)
}
