% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDataSetList.R
\name{Plot.cumulative_difference_plot}
\alias{Plot.cumulative_difference_plot}
\title{Plot the cumulative difference plot given a DataSetList and a runtime.}
\usage{
Plot.cumulative_difference_plot(dsList, runtime)
}
\arguments{
\item{dsList}{A DataSetList (should consist of only one function and dimension and two algorithms).}

\item{runtime}{The target runtime}
}
\value{
A cumulative difference plot.
}
\description{
Plot the cumulative difference plot given a DataSetList and a runtime.
}
\examples{
dsl
dsl_sub <- subset(dsl, funcId == 1)
runtime <- 15
Plot.cumulative_difference_plot(dsl_sub, runtime)
}
