% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDataSetList.R
\name{Plot.FV.ECDF_AUC}
\alias{Plot.FV.ECDF_AUC}
\alias{Plot.FV.ECDF_AUC.DataSetList}
\title{Radarplot of the area under the aggregated ECDF-curve of a DataSetList.}
\usage{
Plot.FV.ECDF_AUC(dsList, rt_min = NULL, rt_max = NULL, rt_step = NULL)

\method{Plot.FV.ECDF_AUC}{DataSetList}(dsList, rt_min = NULL, rt_max = NULL, rt_step = NULL)
}
\arguments{
\item{dsList}{A DataSetList (should consist of only one function and dimension).}

\item{rt_min}{The starting runtime}

\item{rt_max}{The final runtime}

\item{rt_step}{The spacing between starting and final runtimes}
}
\value{
A radarplot of the area under the aggregated ECDF-curve of the DataSetList
}
\description{
Radarplot of the area under the aggregated ECDF-curve of a DataSetList.
}
\examples{
\dontshow{data.table::setDTthreads(1)}
Plot.FV.ECDF_AUC(subset(dsl, funcId == 1))
}
