% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDataSetList.R
\name{Plot.FV.ECDF_Per_Target}
\alias{Plot.FV.ECDF_Per_Target}
\alias{Plot.FV.ECDF_Per_Target.DataSetList}
\title{Plot the empirical cumulative distriburtion as a function of the target values of
a DataSetList at certain target runtimes}
\usage{
Plot.FV.ECDF_Per_Target(dsList, runtimes, scale.xlog = F)

\method{Plot.FV.ECDF_Per_Target}{DataSetList}(dsList, runtimes,
  scale.xlog = F)
}
\arguments{
\item{dsList}{A DataSetList (should consist of only one function and dimension).}

\item{runtimes}{The target runtimes}

\item{scale.xlog}{Whether or not to scale the x-axis logaritmically}
}
\value{
A plot of the empirical cumulative distriburtion as a function of
the fucntion values of the DataSetList at the target runtimes
}
\description{
Plot the empirical cumulative distriburtion as a function of the target values of
a DataSetList at certain target runtimes
}
\examples{
Plot.FV.ECDF_Per_Target(subset(dsl, funcId == 1), 10)
}
