% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDataSetList.R
\name{Plot.FV.PDF}
\alias{Plot.FV.PDF}
\alias{Plot.FV.PDF.DataSetList}
\title{Plot probablity density function of the function values of a DataSetList at
a certain target runtime}
\usage{
Plot.FV.PDF(dsList, runtime, show.sample = F, scale.ylog = F)

\method{Plot.FV.PDF}{DataSetList}(dsList, runtime, show.sample = F, scale.ylog = F)
}
\arguments{
\item{dsList}{A DataSetList (should consist of only one function and dimension).}

\item{runtime}{The target runtime}

\item{show.sample}{Whether or not to show the individual function value samples}

\item{scale.ylog}{Whether or not to scale the y-axis logaritmically}
}
\value{
A plot of the probablity density function of the runtimes at a the
        target function value of the DataSetList
}
\description{
Plot probablity density function of the function values of a DataSetList at
a certain target runtime
}
\examples{
Plot.FV.PDF(subset(dsl, funcId == 1), 100)
}
