% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDataSetList.R
\name{Plot.Performviz}
\alias{Plot.Performviz}
\title{Create the PerformViz plot}
\usage{
Plot.Performviz(DSC_rank_result)
}
\arguments{
\item{DSC_rank_result}{The result from a call to DSCtool rank service (`get_dsc_rank`)}
}
\value{
A performviz plot
}
\description{
From the paper:
}
\examples{
\dontrun{
Plot.Performviz(get_dsc_rank(dsl))
}
}
