% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDataSetList.R
\name{Plot.RT.ECDF_Single_Func}
\alias{Plot.RT.ECDF_Single_Func}
\alias{Plot.RT.ECDF_Single_Func.DataSetList}
\title{Plot the aggregated empirical cumulative distriburtion as a function of the running times of
a DataSetList. Aggregated over multiple functions or dimensions.}
\usage{
Plot.RT.ECDF_Single_Func(dsList, targets = NULL)

\method{Plot.RT.ECDF_Single_Func}{DataSetList}(dsList, targets = NULL)
}
\arguments{
\item{dsList}{A DataSetList.}

\item{targets}{The target function values. Specified in a data.frame, as can be generated
by the function 'get_default_ECDF_targets'}
}
\value{
A plot of the empirical cumulative distriburtion as a function of
the running times of the DataSetList
}
\description{
Plot the aggregated empirical cumulative distriburtion as a function of the running times of
a DataSetList. Aggregated over multiple functions or dimensions.
}
\examples{
Plot.RT.ECDF_Single_Func(subset(dsl, funcId == 1))
}
