% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDataSetList.R
\name{plot_eaf_differences}
\alias{plot_eaf_differences}
\title{Create EAF-difference contour plots}
\usage{
plot_eaf_differences(matrices, scale.xlog = T, scale.ylog = F)
}
\arguments{
\item{matrices}{The dataframes containing the data to plot. This should come from `generate_data.EAF_diff_Approximate`}

\item{scale.xlog}{Logarithmic scaling of x-axis}

\item{scale.ylog}{Logarithmic scaling of y-axis}

\item{...}{Additional parameters for the add_trace function}
}
\value{
EAF difference plots
}
\description{
Create EAF-difference contour plots
}
\examples{
\dontrun{
plot_eaf_differences(generate_data.EAF_diff_Approximate(subset(dsl, funcId == 1), 1, 50, 1, 16))
}
}
