% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDataSetList.R
\name{plot_ERT_AGGR}
\alias{plot_ERT_AGGR}
\title{Plot ERT-based comparison over multiple functions or dimensions}
\usage{
plot_ERT_AGGR(dsList, ...)
}
\arguments{
\item{dsList}{A DataSetList (should consist of only one function and dimension).}

\item{...}{Arguments passed to other methods}
}
\value{
A plot of ERT-based comparison on the provided functions or dimensions of the DataSetList
}
\description{
Plot ERT-based comparison over multiple functions or dimensions
}
