% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDataSetList.R
\name{plot_RT_all_fcts.DataSetList}
\alias{plot_RT_all_fcts.DataSetList}
\title{Plot ERT-plots for multiple functions or dimensions}
\usage{
\method{plot_RT_all_fcts}{DataSetList}(dsList, plot_mode = "subplot",
  scale.xlog = F, scale.ylog = F, scale.reverse = F,
  aggr_on = "funcId", ...)
}
\arguments{
\item{dsList}{A DataSetList (should consist of only one function OR dimension).}

\item{plot_mode}{How the plots should be created. Can be 'subplot' for one plot
per function/dimension or 'overlay' for on plot containing all functions/dimensions}

\item{scale.xlog}{Whether or not to scale the x-axis logaritmically}

\item{scale.ylog}{Whether or not to scale the y-axis logaritmically}

\item{scale.reverse}{Wheter or not to reverse the x-axis (when using minimization)}

\item{aggr_on}{Whether to create a plot for each function ('funcId') or dimension ('DIM')}

\item{...}{Arguments passed to other methods}
}
\value{
A plot of ERT-values of the DataSetList
}
\description{
Plot ERT-plots for multiple functions or dimensions
}
