% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDataSetList.R
\name{plot_FCE_AGGR}
\alias{plot_FCE_AGGR}
\title{Plot expected function value-based comparison over multiple functions or dimensions}
\usage{
plot_FCE_AGGR(dsList, ...)
}
\arguments{
\item{dsList}{A DataSetList (should consist of only one function OR dimension).}

\item{...}{Arguments passed to other methods}
}
\value{
A plot of expected function value-based comparison on the provided functions
 or dimensions of the DataSetList
}
\description{
Plot expected function value-based comparison over multiple functions or dimensions
}
