% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDataSetList.R
\name{plot_FCE_ECDF_PER_TARGET.DataSetList}
\alias{plot_FCE_ECDF_PER_TARGET.DataSetList}
\title{Plot the empirical cumulative distriburtion as a function of the target values of
a DataSetList at certain target runtimes}
\usage{
\method{plot_FCE_ECDF_PER_TARGET}{DataSetList}(dsList, runtimes,
  scale.xlog = F, ...)
}
\arguments{
\item{dsList}{A DataSetList (should consist of only one function and dimension).}

\item{runtimes}{The target runtimes}

\item{scale.xlog}{Whether or not to scale the x-axis logaritmically}

\item{...}{Arguments passed to other methods}
}
\value{
A plot of the empirical cumulative distriburtion as a function of
the fucntion values of the DataSetList at the target runtimes
}
\description{
Plot the empirical cumulative distriburtion as a function of the target values of
a DataSetList at certain target runtimes
}
