% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDataSetList.R
\name{plot_FCE_ECDF_PER_TARGET}
\alias{plot_FCE_ECDF_PER_TARGET}
\title{Plot the empirical cumulative distriburtion as a function of the target values of
a DataSetList at certain target runtimes}
\usage{
plot_FCE_ECDF_PER_TARGET(dsList, ...)
}
\arguments{
\item{dsList}{A DataSetList (should consist of only one function and dimension).}

\item{...}{Arguments passed to other methods}
}
\value{
A plot of the empirical cumulative distriburtion as a function of
the fucntion values of the DataSetList at the target runtimes
}
\description{
Plot the empirical cumulative distriburtion as a function of the target values of
a DataSetList at certain target runtimes
}
