% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDataSetList.R
\name{plot_FCE_MULTI}
\alias{plot_FCE_MULTI}
\title{Plot expected function value-plots for multiple functions or dimensions}
\usage{
plot_FCE_MULTI(dsList, ...)
}
\arguments{
\item{dsList}{A DataSetList (should consist of only one function and dimension).}

\item{...}{Arguments passed to other methods}
}
\value{
A plot of expected function values of the DataSetList
}
\description{
Plot expected function value-plots for multiple functions or dimensions
}
