% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDataSetList.R
\name{plot_FV_line}
\alias{plot_FV_line}
\title{Plot lineplot of the expected function values of a DataSetList}
\usage{
plot_FV_line(dsList, ...)
}
\arguments{
\item{dsList}{A DataSetList (should consist of only one function and dimension).}

\item{...}{Arguments passed to other methods}
}
\value{
A plot of ERT-values of the DataSetList
}
\description{
Plot lineplot of the expected function values of a DataSetList
}
