% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDataSetList.R
\name{plot_FV_PDF}
\alias{plot_FV_PDF}
\title{Plot probablity density function of the function values of a DataSetList at
a certain target runtime}
\usage{
plot_FV_PDF(dsList, ...)
}
\arguments{
\item{dsList}{A DataSetList (should consist of only one function and dimension).}

\item{...}{Arguments passed to other methods}
}
\value{
A plot of the probablity density function of the runtimes at a the
        target function value of the DataSetList
}
\description{
Plot probablity density function of the function values of a DataSetList at
a certain target runtime
}
