% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDataSetList.R
\name{plot_FV_RT_surface}
\alias{plot_FV_RT_surface}
\title{Plot surfaceplot of the expected function values of a DataSetList}
\usage{
plot_FV_RT_surface(dsList, Fstart = NULL, Fstop = NULL,
  RTstart = NULL, RTstop = NULL, surfaces = NULL)
}
\arguments{
\item{dsList}{A DataSetList (should consist of only one function and dimension).}

\item{RTstart}{The starting runtime value.}

\item{RTstop}{The final runtime value.}
}
\value{
A 3dimentional plot of Fixed-Target, Fixed-Budget intersections of the DataSetList
}
\description{
Plot surfaceplot of the expected function values of a DataSetList
}
