% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDataSetList.R
\name{plot_PAR_Line}
\alias{plot_PAR_Line}
\title{Plot the parameter values recorded in a DataSetList}
\usage{
plot_PAR_Line(dsList, ...)
}
\arguments{
\item{dsList}{A DataSetList (should consist of only one function and dimension).}

\item{...}{Arguments passed to other methods}
}
\value{
A plot of for every recorded parameter in the DataSetList
}
\description{
Plot the parameter values recorded in a DataSetList
}
