% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDataSetList.R
\name{plot_RT_all_fcts}
\alias{plot_RT_all_fcts}
\title{Plot ERT-plots for multiple functions or dimensions}
\usage{
plot_RT_all_fcts(dsList, ...)
}
\arguments{
\item{dsList}{A DataSetList (should consist of only one function and dimension).}

\item{...}{Arguments passed to other methods}
}
\value{
A plot of ERT-values of the DataSetList
}
\description{
Plot ERT-plots for multiple functions or dimensions
}
