% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDataSetList.R
\name{plot_RT_AUC}
\alias{plot_RT_AUC}
\title{Radarplot of the area under the aggregated ECDF-curve of a DataSetList.}
\usage{
plot_RT_AUC(dsList, ...)
}
\arguments{
\item{dsList}{A DataSetList (should consist of only one function and dimension).}

\item{...}{Arguments passed to other methods}
}
\value{
A radarplot of the area under the aggregated ECDF-curve of the DataSetList
}
\description{
Radarplot of the area under the aggregated ECDF-curve of a DataSetList.
}
