% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDataSetList.R
\name{plot_RT_ECDF}
\alias{plot_RT_ECDF}
\title{Plot the empirical cumulative distriburtion as a function of the running times of
a DataSetList at certain target function values}
\usage{
plot_RT_ECDF(dsList, ...)
}
\arguments{
\item{dsList}{A DataSetList (should consist of only one function and dimension).}

\item{...}{Arguments passed to other methods}
}
\value{
A plot of the empirical cumulative distriburtion as a function of
the running times of the DataSetList at the target function values
}
\description{
Plot the empirical cumulative distriburtion as a function of the running times of
a DataSetList at certain target function values
}
