% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDataSetList.R
\name{plot_RT_ECDF_AGGR}
\alias{plot_RT_ECDF_AGGR}
\title{Plot the aggregated empirical cumulative distriburtion as a function of the running times of
a DataSetList.}
\usage{
plot_RT_ECDF_AGGR(dsList, ...)
}
\arguments{
\item{dsList}{A DataSetList (should consist of only one function and dimension).}

\item{...}{Arguments passed to other methods}
}
\value{
A plot of the empirical cumulative distriburtion as a function of
the running times of the DataSetList
}
\description{
Plot the aggregated empirical cumulative distriburtion as a function of the running times of
a DataSetList.
}
