% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDataSetList.R
\name{plot_RT_ECDF_MULTI.DataSetList}
\alias{plot_RT_ECDF_MULTI.DataSetList}
\title{Plot the aggregated empirical cumulative distriburtion as a function of the running times of
a DataSetList. Aggregated over multiple functions or dimensions.}
\usage{
\method{plot_RT_ECDF_MULTI}{DataSetList}(dsList, targets = NULL, ...)
}
\arguments{
\item{dsList}{A DataSetList.}

\item{targets}{The target function values. Specified in a data.frame, as can be generated
by the function 'get_default_ECDF_targets'}

\item{...}{Arguments passed to other methods}
}
\value{
A plot of the empirical cumulative distriburtion as a function of
the running times of the DataSetList
}
\description{
Plot the aggregated empirical cumulative distriburtion as a function of the running times of
a DataSetList. Aggregated over multiple functions or dimensions.
}
