% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{read_COCO_dat2}
\alias{read_COCO_dat2}
\title{read COCO '.dat'-like file directly in R}
\usage{
read_COCO_dat2(fname, subsampling = FALSE)
}
\arguments{
\item{fname}{The path to the .dat file}

\item{subsampling}{Whether to subsample the data or not}
}
\value{
A list of data.frames
}
\description{
read COCO '.dat'-like file directly in R
}
\keyword{internal}
