% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{read_dat}
\alias{read_dat}
\title{Read IOHProfiler *.dat files}
\usage{
read_dat(fname, subsampling = FALSE)
}
\arguments{
\item{fname}{The path to the .dat file}

\item{subsampling}{Whether to subsample the data or not}
}
\value{
A list of data.frames
}
\description{
Read IOHProfiler *.dat files
}
