% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{read_dir}
\alias{read_dir}
\title{Read all the data file in a folfer}
\usage{
read_dir(path, verbose = T, print_fun = NULL, maximization = TRUE,
  format = IOHprofiler, subsampling = FALSE)
}
\arguments{
\item{path}{Path to the data files. Will look for all .info-files in this directory and use
the corresponding datafiles to create the DataSetList}

\item{verbose}{Logical.}

\item{print_fun}{Function used to print output when in verbose mode}

\item{maximization}{Logical. Whether the underlying optimization algorithm performs a maximization?}

\item{format}{A character. The format of data source, either 'IOHProfiler', 'COCO' or 'TWO_COL"}

\item{subsampling}{Logical. Whether *.cdat files are subsampled?}
}
\value{
A DataSetList object
}
\description{
Read all the data file in a folfer
}
