% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{read_IndexFile}
\alias{read_IndexFile}
\title{Read .info files and extract information}
\usage{
read_IndexFile(fname)
}
\arguments{
\item{fname}{The path to the .info file}
}
\value{
The data contained in the .info file
}
\description{
Read .info files and extract information
}
\examples{
path <- system.file("extdata", "ONE_PLUS_LAMDA_EA", package="IOHanalyzer")
info <- read_IndexFile(file.path(path,"IOHprofiler_f1_i1.info"))
}
