% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{read_IndexFile_BIOBJ_COCO}
\alias{read_IndexFile_BIOBJ_COCO}
\title{Read Biobjective COCO-based .info files and extract information}
\usage{
read_IndexFile_BIOBJ_COCO(fname)
}
\arguments{
\item{fname}{The path to the .info file}
}
\value{
The data contained in the .info file
}
\description{
Read Biobjective COCO-based .info files and extract information
}
\keyword{internal}
