% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{read_IndexFile_IOH}
\alias{read_IndexFile_IOH}
\title{Read IOHprofiler-based .info files and extract information}
\usage{
read_IndexFile_IOH(fname)
}
\arguments{
\item{fname}{The path to the .info file}
}
\value{
The data contained in the .info file
}
\description{
Read IOHprofiler-based .info files and extract information
}
