% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{read_IOH_v1plus}
\alias{read_IOH_v1plus}
\title{Read Nevergrad data}
\usage{
read_IOH_v1plus(info)
}
\arguments{
\item{info}{A List containing all meta-data about the dataset to create}
}
\value{
The DataSetList extracted from the .csv file provided
}
\description{
Read .csv files in arbitrary format
}
