% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{read_pure_csv}
\alias{read_pure_csv}
\title{Read Nevergrad data}
\usage{
read_pure_csv(
  path,
  neval_name,
  fval_name,
  fname_name,
  algname_name,
  dim_name,
  run_name,
  maximization = F,
  static_attrs = NULL
)
}
\arguments{
\item{path}{The path to the .csv file}

\item{neval_name}{The name of the column to use for the evaluation count.
If NULL, will be assumed to be sequential}

\item{fval_name}{The name of the column to use for the function values}

\item{fname_name}{The name of the column to use for the function name}

\item{algname_name}{The name of the column to use for the algorithm name}

\item{dim_name}{The name of the column to use for the dimension}

\item{run_name}{The name of the column to use for the run number}

\item{maximization}{Boolean indicating whether the data is resulting from maximization or minimization}

\item{static_attrs}{Named list containing the static values for missing columns.
When a parameter is not present in the csv file, its name-parameter should
be set to NULL, and the static value should be added to this static_attrs list.}
}
\value{
The DataSetList extracted from the .csv file provided
}
\description{
Read .csv files in arbitrary format
}
