% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{register_DSC}
\alias{register_DSC}
\title{Register an account to the DSCtool API}
\usage{
register_DSC(name, username, affiliation, email, password = NULL)
}
\arguments{
\item{name}{Your name}

\item{username}{A usename to be identified with. Will be stored on keyring under 'DSCtool_name'}

\item{affiliation}{Your affiliation (university / company)}

\item{email}{Your email adress}

\item{password}{The password to use. If NULL, this will be generated at random. 
Will be stored on keyring under 'DSCtool'}
}
\description{
This uses the keyring package to store and load credentials. 
If you already have an account, please call `add_DSC_credentials` instead
}
\examples{
\dontrun{ 
register_DSC('John Doe', 'jdoe', 'Sample University', "j.doe.sample.com")
}
}
