% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{save_plotly}
\alias{save_plotly}
\title{Save plotly figure in multiple format}
\usage{
save_plotly(p, file, format = "svg", ...)
}
\arguments{
\item{p}{plotly object. The plot to be saved}

\item{file}{String. The name of the figure file}

\item{format}{String. The format of the figure: 'svg', 'pdf', 'eps', 'png' are supported}

\item{...}{Additional arguments for orca}
}
\description{
Save plotly figure in multiple format
}
