% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{save_table}
\alias{save_table}
\title{Save DataTable in multiple formats}
\usage{
save_table(df, file, format = NULL)
}
\arguments{
\item{df}{The DataTable to store}

\item{file}{String. The name of the figure file, with the extension of the required file-format}

\item{format}{Optional, string. Overwrites the extension of the `file` parameter. If not specified while
file does not have an extension, it defaults to csv}
}
\description{
Save DataTable in multiple formats
}
\examples{
df <- generate_data.Single_Function(subset(dsl, funcId == 1), which = 'by_RT')
save_table(df, tempfile(fileext = ".md"))
}
