% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{scan_index_file}
\alias{scan_index_file}
\title{Scan *.info files for IOHProfiler or COCO}
\usage{
scan_index_file(folder)
}
\arguments{
\item{folder}{The folder containing the .info or .json files}
}
\value{
The paths to all found .info and .json-files
}
\description{
Scan *.info files for IOHProfiler or COCO
}
\note{
This automatically filetrs our files of size 0
}
\examples{
path <- system.file("extdata", "ONE_PLUS_LAMDA_EA", package="IOHanalyzer")
scan_index_file(path)
}
