% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{scan_indexFile}
\alias{scan_indexFile}
\title{Scan *.info files for IOHProfiler or COCO}
\usage{
scan_indexFile(folder)
}
\arguments{
\item{folder}{The folder containing the .info files}
}
\value{
The paths to all found .info-files
}
\description{
Scan *.info files for IOHProfiler or COCO
}
