% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{set_color_scheme}
\alias{set_color_scheme}
\title{Set the colorScheme of the IOHanalyzer plots}
\usage{
set_color_scheme(schemename, ids, path = NULL)
}
\arguments{
\item{schemename}{Three default colorschemes are implemented:
\itemize{
\item Default
\item Variant 1
\item Variant 2
\item Variant 3
}
And it is also possible to select "Custom", which allows uploading of a custom set of colors}

\item{ids}{The names of the algorithms (or custom ids, see `change_id`) for which to set the colors}

\item{path}{The path to the file containing the colors to use. Only used if
schemename is "Custom"}
}
\description{
Set the colorScheme of the IOHanalyzer plots
}
\examples{
set_color_scheme("Default", get_algId(dsl))
}
