% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{set_colorScheme}
\alias{set_colorScheme}
\title{Set the colorScheme of the IOHanalyzer plots}
\usage{
set_colorScheme(schemename, path = NULL)
}
\arguments{
\item{schemename}{Three default colorschemes are implemented:
\itemize{
\item Default
\item Variant 1
\item Variant 2
}
And it is also possible to select "Custom", which allows uploading of a custom set of colors}

\item{path}{The path to the file containing the colors to use. Only used if 
schemename is "Custom"}
}
\description{
Set the colorScheme of the IOHanalyzer plots
}
\examples{
set_colorScheme("Default")
}
