% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{set_DSC_credentials}
\alias{set_DSC_credentials}
\title{Register an account to the DSCtool API}
\usage{
set_DSC_credentials(username, password)
}
\arguments{
\item{username}{The usename you use on DSCtool. Will be stored on keyring under 'DSCtool_name'}

\item{password}{The password you use on DSCtool. Will be stored on keyring under 'DSCtool'}
}
\description{
This uses the keyring package to store and load credentials. 
If you already have an account, please call `add_DSC_credentials` instead
}
\examples{
\dontrun{set_DSC_credentials('jdoe', 'monkey123')}
}
