% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOHanalyzer.R
\name{set_property}
\alias{set_property}
\title{Set the value of some global IOHanalyzer properties}
\usage{
set_property(prop_name, value, ...)
}
\arguments{
\item{prop_name}{The property to set. Current options are:
\itemize{
\item 'probs': The probabilities in the displayed quantiles (RT and FV summaries). Needs a vector of probabilities
\item 'max_samples': The maximum number of samples to generate for each algorithm. Needs a single integer
}}

\item{value}{The value to set the property to}

\item{...}{Addition arguments}
}
\description{
Set the value of some global IOHanalyzer properties
}
\examples{
set_property("probs",c(0.1,0.25,0.5,0.75,0.9))
}
