% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{SP}
\alias{SP}
\title{Estimator 'SP' for the Expected Running Time (ERT)}
\usage{
SP(data, max_runtime)
}
\arguments{
\item{data}{A dataframe or matrix. Each row stores the runtime sample points from
several runs}

\item{max_runtime}{The budget to use for calculating ERT. If this is a vector, the largest value is taken.
Using this as a vector is being deprecated, and will be removed in a future update}
}
\value{
A list containing ERTs, number of succesfull runs and the succes rate
}
\description{
Estimator 'SP' for the Expected Running Time (ERT)
}
\examples{
SP(dsl[[1]]$RT, max(dsl[[1]]$RT))
}
