% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{[.DataSetList}
\alias{[.DataSetList}
\title{S3 extraction function for DataSetList}
\usage{
\method{[}{DataSetList}(x, i, drop = FALSE)
}
\arguments{
\item{x}{The DataSetList to use}

\item{i}{The indices to extract}

\item{drop}{Currently unused parameter}
}
\value{
The DataSetList of the DataSets at indices i of DataSetList x
}
\description{
S3 extraction function for DataSetList
}
\examples{
dsl[c(1, 3)]
}
