% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\name{subset.DataSet}
\alias{subset.DataSet}
\title{S3 subset function for DataSet}
\usage{
\method{subset}{DataSet}(x, mask, ...)
}
\arguments{
\item{x}{The DataSet from which to get a subset}

\item{mask}{The mask to use when subsetting. The length should be equal to the number of runs present in the 
provided dataset object x.}

\item{...}{Arguments passed to underlying subset method (not yet supported)}
}
\value{
A new DataSet
}
\description{
Subset for DataSets. Based on the provided mask, the relevant data is taken from the given DataSet
and turned into a new DataSet object.
}
\examples{
subset(ds, [0,1,1,1])
}
