% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{subset.DataSetList}
\alias{subset.DataSetList}
\title{Filter a DataSetList by some criteria}
\usage{
\method{subset}{DataSetList}(x, ...)
}
\arguments{
\item{x}{The DataSetList}

\item{...}{The conditions to filter on. Can be any expression which assigns True or False
to a DataSet object, such as DIM == 625 or funcId == 2. Multiple conditions should be seperated by commas and will
be treated as AND. Support for custom conditions using AND and OR are not supported as of version 0.1.6.0}
}
\value{
The filtered DataSetList
}
\description{
Filter a DataSetList by some criteria
}
\examples{
subset(dsl, funcId == 1)
}
