% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{summary.DataSetList}
\alias{summary.DataSetList}
\title{S3 summary function for DataSetList}
\usage{
\method{summary}{DataSetList}(object, ...)
}
\arguments{
\item{object}{The DataSetList to print}

\item{...}{Arguments for underlying summary function?}
}
\description{
Prints the Function ID, Dimension, Algorithm Id, datafile location and comment for every
DataSet in the DataSetList
}
\examples{
summary(dsl)
}
