% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{align_runtime}
\alias{align_runtime}
\title{Align data by runtimes}
\usage{
align_runtime(data, format = IOHprofiler)
}
\arguments{
\item{data}{The data to align}

\item{format}{Whether the data is form IOHprofiler or COCO.}
}
\value{
Data aligned by runtime
}
\description{
Align data by runtimes
}
