% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{arrange}
\alias{arrange}
\alias{arrange.DataSetList}
\title{S3 sort function for DataSetList}
\usage{
arrange(dsl, ...)

\method{arrange}{DataSetList}(dsl, ...)
}
\arguments{
\item{dsl}{The DataSetList to sort}

\item{...}{attribute by which `dsl` is sorted. Multiple attributes can be specified.}
}
\description{
Sorts a DataSetList based on the custom specified attributes ('algId', 'DIM' or 'funcId'). 
Default is as ascending, can be made descending by adding a - in front of the attribute.
Sorting accross multiple attributes is supported, in the order they are specified.
}
\examples{
arrange(dsl, DIM, -funcId, algId) 
}
