% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{check_dsc_configured}
\alias{check_dsc_configured}
\title{Verify that the credentials for DSCtool have been set}
\usage{
check_dsc_configured()
}
\description{
This uses the keyring package to store and load credentials. 
If you already have an account, please call `set_DSC_credentials`
with the corresponding username and password. 
If you don't have an account, you can register for one using `register_DSC`
}
\examples{
check_dsc_configured()
}
