% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{check_format}
\alias{check_format}
\title{Check the format of data}
\usage{
check_format(path)
}
\arguments{
\item{path}{The path to the folder to check}
}
\value{
The format of the data in the given folder. Either 'COCO' or 'IOHprofiler'.
}
\description{
Throws a warning when multiple formats are found in the same folder.
}
