% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{generate_data.CDP}
\alias{generate_data.CDP}
\title{Generate data for the cumulative difference plot.}
\usage{
generate_data.CDP(dsList, runtime)
}
\arguments{
\item{dsList}{The DataSetList object.
Note that the `cumulative_difference_plot` can only compare two algorithms
in a single problem of dimension one.}
}
\value{
A dataframe with the data to generate the cumulative difference plot.
}
\description{
This function generates a dataframe that can be used to generate
the `cumulative_difference_plot`-function
}
\examples{

generate_data.CDP(subset(dsl, funcId == 1), 25)
}
