% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{generate_data.EAF}
\alias{generate_data.EAF}
\title{Generate dataframe consisting of the levelsets of the EAF}
\usage{
generate_data.EAF(dsList, n_sets = 11)
}
\arguments{
\item{dsList}{The DataSetList object}

\item{n_sets}{The number of level sets to calculate}
}
\description{
This function generates a dataframe which can be easily plotted using the `plot_eaf_data`-function
}
\examples{
generate_data.EAF(subset(dsl, funcId == 1))
}
