% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{get_algId}
\alias{get_algId}
\title{Get all algorithm ids present in a DataSetList}
\usage{
get_algId(dsList)
}
\arguments{
\item{dsList}{The DataSetLsit}
}
\value{
A sorted list of all unique algorithm ids which occur in the DataSetList
}
\description{
Get all algorithm ids present in a DataSetList
}
\examples{
get_algId(dsl)
}
