% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R, R/DataSetList.R
\name{get_ERT}
\alias{get_ERT}
\alias{get_ERT.DataSet}
\alias{get_ERT.DataSetList}
\title{Get Expected RunTime}
\usage{
get_ERT(ds, ftarget, budget, ...)

\method{get_ERT}{DataSet}(ds, ftarget, budget = NULL, ...)

\method{get_ERT}{DataSetList}(ds, ftarget, budget = NULL,
  algorithm = "all", ...)
}
\arguments{
\item{ds}{A DataSet or DataSetList object}

\item{ftarget}{The function target(s) for which to get the ERT}

\item{budget}{Optional; overwrites the budget found in ds for ERT-calculation}

\item{...}{Arguments passed to other methods}

\item{algorithm}{DEPRECATED, will be removed in next release. Which algorithms in the DataSetList to consider.}
}
\value{
A data.table containing the runtime samples for each provided target
function value
}
\description{
Get Expected RunTime
}
\examples{
get_ERT(dsl, 14)
get_ERT(dsl[[1]], 14)
}
