% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{get_funvals}
\alias{get_funvals}
\title{Get all function values present in a DataSetList}
\usage{
get_funvals(dsList)
}
\arguments{
\item{dsList}{The DataSetLsit}
}
\value{
A list matrices of all function values which occur in the DataSetList
}
\description{
Get all function values present in a DataSetList
}
