% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R, R/DataSetList.R
\name{get_FV_overview}
\alias{get_FV_overview}
\alias{get_FV_overview.DataSet}
\alias{get_FV_overview.DataSetList}
\title{Get Function Value condensed overview}
\usage{
get_FV_overview(ds, ...)

\method{get_FV_overview}{DataSet}(ds, ...)

\method{get_FV_overview}{DataSetList}(ds, algorithm = "all", ...)
}
\arguments{
\item{ds}{A DataSet or DataSetList object}

\item{...}{Arguments passed to other methods}

\item{algorithm}{Which algorithms in the DataSetList to consider.}
}
\value{
A data.table containing the algorithm ID, best, worst and mean reached function
values, the number of runs and available budget for the DataSet
}
\description{
Get Function Value condensed overview
}
\examples{
get_FV_overview(dsl)
get_FV_overview(dsl[[1]])
get_FV_overview(dsl, algorithm = "(1+1)_greedy_hill_climber_1" )
}
