% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\name{get_FV_sample}
\alias{get_FV_sample}
\title{Get Funtion Value Samples}
\usage{
get_FV_sample(ds, ...)
}
\arguments{
\item{ds}{A DataSet or DataSetList object}

\item{...}{Arguments passed to other methods}
}
\value{
A data.table containing the function value samples for each provided
target runtime
}
\description{
Get Funtion Value Samples
}
