% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R, R/DataSetList.R
\name{get_FV_summary}
\alias{get_FV_summary}
\alias{get_FV_summary.DataSet}
\alias{get_FV_summary.DataSetList}
\title{Get Function Value Summary}
\usage{
get_FV_summary(ds, ...)

\method{get_FV_summary}{DataSet}(ds, runtime, include_geom_mean = F, include_limits = F, ...)

\method{get_FV_summary}{DataSetList}(
  ds,
  runtime,
  algorithm = "all",
  include_geom_mean = F,
  include_limits = F,
  ...
)
}
\arguments{
\item{ds}{A DataSet or DataSetList object}

\item{...}{Arguments passed to other methods}

\item{runtime}{A Numerical vector. Runtimes at which function values are reached}

\item{include_geom_mean}{Boolean to indicate whether to include the geometric mean.
Only works in fixed_budget mode. Negative values cause NaN, zeros cause output to be completely 0. Defaults to False.}

\item{include_limits}{Boolean to indicate whether to include the min and max.}

\item{algorithm}{DEPRECATED, will be removed in next release. Which algorithms in the DataSetList to consider.}
}
\value{
A data.table containing the function value statistics for each provided
target runtime value
}
\description{
Get Function Value Summary
}
\examples{
get_FV_summary(dsl, 100)
get_FV_summary(dsl[[1]], 100)
}
