% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R, R/DataSetList.R
\name{get_FV_summary}
\alias{get_FV_summary}
\alias{get_FV_summary.DataSet}
\alias{get_FV_summary.DataSetList}
\title{Get Function Value Summary}
\usage{
get_FV_summary(ds, ...)

\method{get_FV_summary}{DataSet}(ds, runtime, ...)

\method{get_FV_summary}{DataSetList}(ds, runtime, algorithm = "all", ...)
}
\arguments{
\item{ds}{A DataSet or DataSetList object}

\item{...}{Arguments passed to other methods}

\item{runtime}{A Numerical vector. Runtimes at which function values are reached}

\item{algorithm}{DEPRECATED, will be removed in next release. Which algorithms in the DataSetList to consider.}
}
\value{
A data.table containing the function value statistics for each provided
target runtime value
}
\description{
Get Function Value Summary
}
\examples{
get_FV_summary(dsl, 100)
get_FV_summary(dsl[[1]], 100)
}
