% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R, R/DataSetList.R
\name{get_id}
\alias{get_id}
\alias{get_id.DataSet}
\alias{get_id.DataSetList}
\title{Get condensed overview of datasets}
\usage{
get_id(ds, ...)

\method{get_id}{DataSet}(ds, ...)

\method{get_id}{DataSetList}(ds, ...)
}
\arguments{
\item{ds}{The DataSetList}

\item{...}{Arguments passed to other methods}
}
\value{
The list of unique identiefiers present in dsl
}
\description{
Get the unique identifiers for each DataSet in the provided DataSetList
}
\details{
If no unique identifier is set (using `change_id` or done in DataSet construction from 1.6.0 onwards), 
this function falls back on returning the algorith id (from `get_aldId`)to ensure backwards compatibility
}
\examples{
get_id(dsl)
get_id(dsl[[1]])
}
