% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_marg_contrib_ecdf}
\alias{get_marg_contrib_ecdf}
\title{Get the marginal contribution of an algorithm to a portfolio}
\usage{
get_marg_contrib_ecdf(alg, perm, j, dt)
}
\arguments{
\item{alg}{The result from a call to `get_dsc_omnibus`}

\item{perm}{The permutation of algorithms to which is being contributed}

\item{j}{At which point in the permutation the contribution should be measured}

\item{dt}{The datatable in which the raw ecdf-values are stored (see `generate_data.ECDF_raw`)}
}
\description{
Based on the contribution to the ECDF-curve of the VBS of the portfolio
}
\examples{
dt <- generate_data.ECDF_raw(dsl, get_ECDF_targets(dsl))
get_marg_contrib_ecdf(get_algId(dsl)[[1]], get_algId(dsl), 1, dt)
}
