% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\name{get_PAR_sample}
\alias{get_PAR_sample}
\title{Get Parameter Value Samples}
\usage{
get_PAR_sample(ds, ...)
}
\arguments{
\item{ds}{A DataSet or DataSetList object}

\item{...}{Arguments passed to other methods}
}
\value{
A data.table object containing parameter values aligned at each given target value
}
\description{
Get Parameter Value Samples
}
