% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\name{get_PAR_summary}
\alias{get_PAR_summary}
\title{Get Parameter Value Summary}
\usage{
get_PAR_summary(ds, ...)
}
\arguments{
\item{ds}{A DataSet or DataSetList object}

\item{...}{Arguments passed to other methods}
}
\value{
A data.table object containing basic statistics of parameter values aligned at each given target value
}
\description{
Get Parameter Value Summary
}
