% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{get_parId}
\alias{get_parId}
\title{Get all parameter ids present in a DataSetList}
\usage{
get_parId(dsList, which = "by_FV")
}
\arguments{
\item{dsList}{The DataSetList}

\item{which}{A string takes values in `c('by_FV', 'by_RT')`. To choose the parameters aligned
by the running time (RT) or the function value (FV). Note that parameters in each case are
not necessary the same.}
}
\value{
A sorted list of all unique parameter ids which occur in the DataSetList
}
\description{
Get all parameter ids present in a DataSetList
}
\examples{
get_parId(dsl)
}
