% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{get_ParId}
\alias{get_ParId}
\title{Get all parameter ids present in a DataSetList}
\usage{
get_ParId(data)
}
\arguments{
\item{data}{The DataSetLsit}
}
\value{
A sorted list of all unique parameter ids which occur in the DataSetList
}
\description{
Get all parameter ids present in a DataSetList
}
