% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R, R/DataSetList.R
\name{get_RT_overview}
\alias{get_RT_overview}
\alias{get_RT_overview.DataSet}
\alias{get_RT_overview.DataSetList}
\title{Get Runtime Value condensed overview}
\usage{
get_RT_overview(ds, ...)

\method{get_RT_overview}{DataSet}(ds, ...)

\method{get_RT_overview}{DataSetList}(ds, algorithm = "all", ...)
}
\arguments{
\item{ds}{A DataSet or DataSetList object}

\item{...}{Arguments passed to other methods}

\item{algorithm}{DEPRECATED, will be removed in next release. Which algorithms in the DataSetList to consider.}
}
\value{
A data.table containing the algorithm ID, minimum and maximum used evaluations,
number of runs and available budget for the DataSet
}
\description{
Get Runtime Value condensed overview
}
\examples{
get_RT_overview(dsl)
get_RT_overview(dsl[[1]])
}
