% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDataSetList.R
\name{Plot.Stats.Glicko2_Candlestick}
\alias{Plot.Stats.Glicko2_Candlestick}
\alias{Plot.Stats.Glicko2_Candlestick.DataSetList}
\title{Create a candlestick plot of Glicko2-rankings}
\usage{
Plot.Stats.Glicko2_Candlestick(dsList, nr_rounds = 100, glicko2_rank_df = NULL)

\method{Plot.Stats.Glicko2_Candlestick}{DataSetList}(dsList, nr_rounds = 100, glicko2_rank_df = NULL)
}
\arguments{
\item{dsList}{A DataSetList}

\item{nr_rounds}{The number of rounds in the tournament}

\item{glicko2_rank_df}{Optional. Dataframe containing the glicko2 rating to avoid needless recalculation.}
}
\description{
Create a candlestick plot of Glicko2-rankings
}
\examples{
Plot.Stats.Glicko2_Candlestick(dsl, nr_rounds=2)
}
