% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDataSetList.R
\name{Plot.Stats.Significance_Graph}
\alias{Plot.Stats.Significance_Graph}
\alias{Plot.Stats.Significance_Graph.DataSetList}
\title{Plot a network graph showing the statistically different algorithms}
\usage{
Plot.Stats.Significance_Graph(dsList, ftarget, alpha = 0.01,
  bootstrap.size = 30, which = "by_FV")

\method{Plot.Stats.Significance_Graph}{DataSetList}(dsList, ftarget,
  alpha = 0.01, bootstrap.size = 30, which = "by_FV")
}
\arguments{
\item{dsList}{A DataSetList (should consist of only one function and dimension).}

\item{ftarget}{The target function value to use}

\item{alpha}{The cutoff for statistical significance}

\item{bootstrap.size}{The amound of bootstrapped samples used}

\item{which}{Whether to use fixed-target ('by_FV') or fixed-budget ('by_RT') perspective}
}
\value{
A graph showing the statistical significance between algorithms
}
\description{
Plot a network graph showing the statistically different algorithms
}
\examples{
Plot.Stats.Significance_Graph(subset(dsl, funcId == 2), 16)
}
