% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDataSetList.R
\name{plot_FV_HIST}
\alias{plot_FV_HIST}
\title{Plot histograms of the function values of a DataSetList at a certain target runtime}
\usage{
plot_FV_HIST(dsList, ...)
}
\arguments{
\item{dsList}{A DataSetList (should consist of only one function and dimension).}

\item{...}{Arguments passed to other methods}
}
\value{
A plot of the histograms of the function values at a the
        target runtime of the DataSetList
}
\description{
Plot histograms of the function values of a DataSetList at a certain target runtime
}
