% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDataSetList.R
\name{plot_FV_line}
\alias{plot_FV_line}
\alias{plot_FV_line.DataSetList}
\title{Plot lineplot of the expected function values of a DataSetList}
\usage{
plot_FV_line(dsList, RTstart = NULL, RTstop = NULL, show.CI = F,
  show.mean = T, show.median = F, show.runs = F, show.density = 50,
  show.grad = F, show.intensity = 0, show.pareto = F,
  show.optimal = F, backend = "plotly", scale.xlog = F,
  scale.ylog = F, scale.reverse = F)

\method{plot_FV_line}{DataSetList}(dsList, RTstart = NULL,
  RTstop = NULL, show.CI = T, show.mean = F, show.median = F,
  show.runs = F, show.density = 50, show.grad = F,
  show.intensity = 0, show.pareto = F, show.optimal = F,
  backend = "plotly", scale.xlog = F, scale.ylog = F,
  scale.reverse = F)
}
\arguments{
\item{dsList}{A DataSetList (should consist of only one function and dimension).}

\item{RTstart}{The starting runtime value.}

\item{RTstop}{The final runtime value.}

\item{show.CI}{Whether or not to show the standard deviations}

\item{show.mean}{Whether or not to show the mean runtimes}

\item{show.median}{Whether or not to show the median runtimes}

\item{show.runs}{Whether or not to show the individual runs}

\item{show.density}{Percentage of individual runs to show when show.runs is true}

\item{show.grad}{Whether or not to show the run intensity}

\item{show.intensity}{Intensity to use when using show.grad}

\item{show.pareto}{Whether or not to show the pareto-optimal run}

\item{show.optimal}{Whether or not to show the optimal run}

\item{backend}{Which plotting library to use. Can be 'plotly' or 'ggplot2'}

\item{scale.xlog}{Whether or not to scale the x-axis logaritmically}

\item{scale.ylog}{Whether or not to scale the y-axis logaritmically}

\item{scale.reverse}{Wheter or not to reverse the x-axis (when using minimization)}
}
\value{
A plot of ERT-values of the DataSetList
}
\description{
Plot lineplot of the expected function values of a DataSetList
}
\examples{
plot_FV_line(subset(dsl, funcId == 1))
}
