% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDataSetList.R
\name{plot_RT_ECDF}
\alias{plot_RT_ECDF}
\alias{plot_RT_ECDF.DataSetList}
\title{Plot the empirical cumulative distriburtion as a function of the running times of
a DataSetList at certain target function values}
\usage{
plot_RT_ECDF(dsList, ftargets, scale.xlog = F)

\method{plot_RT_ECDF}{DataSetList}(dsList, ftargets, scale.xlog = F)
}
\arguments{
\item{dsList}{A DataSetList (should consist of only one function and dimension).}

\item{ftargets}{The target function values}

\item{scale.xlog}{Whether or not to scale the x-axis logaritmically}
}
\value{
A plot of the empirical cumulative distriburtion as a function of
the running times of the DataSetList at the target function values
}
\description{
Plot the empirical cumulative distriburtion as a function of the running times of
a DataSetList at certain target function values
}
\examples{
plot_RT_ECDF(subset(dsl, funcId == 1), 14)
}
