% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDataSetList.R
\name{plot_RT_PMF}
\alias{plot_RT_PMF}
\title{Plot probablity mass function of the runtimes of a DataSetList at a certain target function value}
\usage{
plot_RT_PMF(dsList, ...)
}
\arguments{
\item{dsList}{A DataSetList (should consist of only one function and dimension).}

\item{...}{Arguments passed to other methods}
}
\value{
A plot of the probablity mass function of the runtimes at a the
        target function value of the DataSetList
}
\description{
Plot probablity mass function of the runtimes of a DataSetList at a certain target function value
}
