% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{read_pure_csv}
\alias{read_pure_csv}
\title{Read Nevergrad data}
\usage{
read_pure_csv(
  path,
  neval_name,
  fval_name,
  fname_name,
  algname_name,
  dim_name,
  run_name,
  maximization = F
)
}
\arguments{
\item{path}{The path to the .csv file}

\item{neval_name}{The name of the column to use for the evaluation count}

\item{fval_name}{The name of the column to use for the function values}

\item{fname_name}{The name of the column to use for the function name}

\item{algname_name}{The name of the column to use for the algorithm name}

\item{dim_name}{The name of the column to use for the dimension}

\item{run_name}{The name of the column to use for the run number}

\item{maximization}{Boolean indicating whether the data is resulting from maximization or minimization}
}
\value{
The DataSetList extracted from the .csv file provided
}
\description{
Read .csv files in arbitrary format
}
