% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runServer.R
\name{runServer}
\alias{runServer}
\title{Create a shiny-server GUI to interactively use the IOHanalyzer}
\usage{
runServer(port = getOption("shiny.port"))
}
\arguments{
\item{port}{Optional; which port the server should be opened at}
}
\description{
Create a shiny-server GUI to interactively use the IOHanalyzer
}
\examples{
\dontrun{
runServer()
}
}
