% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{save_plotly}
\alias{save_plotly}
\title{Save plotly figure in multiple format}
\usage{
save_plotly(p, file, width = NULL, height = NULL, format = NULL, ...)
}
\arguments{
\item{p}{plotly object. The plot to be saved}

\item{file}{String. The name of the figure file}

\item{width}{Optional. Width of the figure}

\item{height}{Optional. Height of the figure}

\item{format}{Deprecated. Optional extra specifier for format}

\item{...}{Additional arguments for orca}
}
\description{
NOTE: This function requires orca to be installed, and for pdf and eps formats
inkscape is also needed.
}
\examples{
\donttest{
p <- Plot.RT.Single_Func(dsl[1])
save_plotly(p, 'example_file.png')
}
}
