% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSetList.R
\name{subset.DataSetList}
\alias{subset.DataSetList}
\title{Filter a DataSetList by some criterium}
\usage{
\method{subset}{DataSetList}(x, ...)
}
\arguments{
\item{x}{The DataSetLsit}

\item{...}{The condition to filter on. Can be any expression which assigns True or False
to a DataSet object, such as DIM==625 or funcId==2}
}
\value{
The filtered DataSetList
}
\description{
Filter a DataSetList by some criterium
}
